requirements.txt
streamlit>=1.40.0
langgraph>=0.2.0
langchain-core>=0.3.0
google-genai>=1.0.0
requests>=2.31.0
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0
